+++ packages/create-svelte/templates/libskeleton/src/app.d.ts
@@ -5,6 +5,7 @@
 // and what to do when importing types
 declare namespace App {
 	// interface Locals {}
+	// interface PageData {}
 	// interface Platform {}
 	// interface PrivateEnv {}
 	// interface PublicEnv {}