+++ packages/kit/src/core/sync/create_manifest_data/index.js
@@ -109,23 +109,25 @@ function create_routes_and_nodes(cwd, config, fallback) {
 
 			segment_map.set(
 				id,
-				segments.filter(Boolean).map((segment) => {
-					/** @type {import('./types').Part[]} */
-					const parts = [];
-					segment.split(/\[(.+?)\]/).map((content, i) => {
-						const dynamic = !!(i % 2);
+				segments
+					.filter((segment) => segment !== '' && affects_path(segment))
+					.map((segment) => {
+						/** @type {import('./types').Part[]} */
+						const parts = [];
+						segment.split(/\[(.+?)\]/).map((content, i) => {
+							const dynamic = !!(i % 2);
 
-						if (!content) return;
+							if (!content) return;
 
-						parts.push({
-							content,
-							dynamic,
-							rest: dynamic && content.startsWith('...'),
-							type: (dynamic && content.split('=')[1]) || null
+							parts.push({
+								content,
+								dynamic,
+								rest: dynamic && content.startsWith('...'),
+								type: (dynamic && content.split('=')[1]) || null
+							});
 						});
-					});
-					return parts;
-				})
+						return parts;
+					})
 			);
 
 			/** @type {import('types').RouteData} */
@@ -180,7 +182,7 @@ function create_routes_and_nodes(cwd, config, fallback) {
 							component: project_relative
 						};
 					} else if (item.is_layout) {
-						if (!route.layout) route.layout = { depth };
+						if (!route.layout) route.layout = { depth, child_pages: [] };
 						route.layout.component = project_relative;
 						if (item.uses_layout !== undefined) route.layout.parent_id = item.uses_layout;
 					} else {
@@ -189,7 +191,7 @@ function create_routes_and_nodes(cwd, config, fallback) {
 						if (item.uses_layout !== undefined) route.leaf.parent_id = item.uses_layout;
 					}
 				} else if (item.is_layout) {
-					if (!route.layout) route.layout = { depth };
+					if (!route.layout) route.layout = { depth, child_pages: [] };
 					route.layout[item.kind] = project_relative;
 				} else if (item.is_page) {
 					if (!route.leaf) route.leaf = { depth };
@@ -223,7 +225,7 @@ function create_routes_and_nodes(cwd, config, fallback) {
 		}
 
 		if (!root.layout?.component) {
-			if (!root.layout) root.layout = { depth: 0 };
+			if (!root.layout) root.layout = { depth: 0, child_pages: [] };
 			root.layout.component = posixify(path.relative(cwd, `${fallback}/layout.svelte`));
 		}
 
@@ -290,6 +292,9 @@ function create_routes_and_nodes(cwd, config, fallback) {
 					}
 
 					if (current_route.layout) {
+						/** @type {import('types').PageNode[]} */ (current_route.layout.child_pages).push(
+							route.leaf
+						);
 						current_node.parent = current_node = current_route.layout;
 						parent_id = current_node.parent_id;
 					} else {