+++ packages/kit/src/runtime/client/parse.js
@@ -10,11 +10,6 @@ export function parse(nodes, dictionary, matchers) {
 	return Object.entries(dictionary).map(([id, [leaf, layouts, errors]]) => {
 		const { pattern, names, types } = parse_route_id(id);
 
-		// whether or not the route uses the server data is
-		// encoded using the ones' complement, to save space
-		const uses_server_data = leaf < 0;
-		if (uses_server_data) leaf = ~leaf;
-
 		const route = {
 			id,
 			/** @param {string} path */
@@ -23,9 +18,8 @@ export function parse(nodes, dictionary, matchers) {
 				if (match) return exec(match, names, types, matchers);
 			},
 			errors: [1, ...(errors || [])].map((n) => nodes[n]),
-			layouts: [0, ...(layouts || [])].map((n) => nodes[n]),
-			leaf: nodes[leaf],
-			uses_server_data
+			layouts: [0, ...(layouts || [])].map(create_loader),
+			leaf: create_loader(leaf)
 		};
 
 		// bit of a hack, but ensures that layout/error node lists are the same
@@ -38,4 +32,16 @@ export function parse(nodes, dictionary, matchers) {
 
 		return route;
 	});
+
+	/**
+	 * @param {number} id
+	 * @returns {[boolean, import('types').CSRPageNodeLoader]}
+	 */
+	function create_loader(id) {
+		// whether or not the route uses the server data is
+		// encoded using the ones' complement, to save space
+		const uses_server_data = id < 0;
+		if (uses_server_data) id = ~id;
+		return [uses_server_data, nodes[id]];
+	}
 }