+++ packages/kit/test/apps/basics/src/routes/errors/+layout.js
@@ -1,0 +1,9 @@
+import { error } from '@sveltejs/kit';
+
+/** @type {import('@sveltejs/kit').Load} */
+export async function load({ fetch, url }) {
+	if (url.pathname.startsWith('/errors/error-in-layout')) {
+		const res = await fetch('/errors/error-in-layout/non-existent');
+		throw error(res.status);
+	}
+}